requirements.txt
streamlit==1.20.0
folium==0.14.0
streamlit_folium==0.11.0
python-dotenv==0.21.1
pandas==1.5.3
geopy==2.3.0
pyproj==3.4.1
shapely==2.0.1
rasterio==1.3.9
scikit-learn==1.2.1
matplotlib==3.6.3
opencv-python==4.7.0.68
Pyrebase4==4.7.1
urllib3==1.26.15
requests-toolbelt==0.10.1
numpy~=1.26.4
requests~=2.29.0
Jinja2~=3.1.3